added size_t type where it was necessary

+++ headers/credit.h
@@ -7,7 +7,7 @@ int digitCount(long int);
 int checkLengthValidity(long int);
 int * getEverySecond(long int);
 int getRestSum(long int);
-int productsDigits(int*,int);
+int productsDigits(int*,size_t);
 int validate(int, int);
 
 int checkAmex(long int);
